DB: Initial commit for Badgerloop member data
Will likely need refactoring when secure SSO
is back up to see what other fields could
be added to the table(s) but this should
be a good start. There's danger here because
if the table-creation SQL script runs it will
drop all existing data. Might be a good idea
to set up a Jenkins job that creates SQL script
backups of all of the data.

Signed-off-by: vkottler <vkottler@wisc.edu>